add raw mode
